HHC-97 Change salary max value Int -> Long

+++ app/src/main/java/ru/practicum/android/diploma/domain/models/Filter.kt
@@ -3,6 +3,6 @@ package ru.practicum.android.diploma.domain.models
 data class Filter(
     val workPlace: Area? = null,
     val industry: Industry? = null,
-    val salary: Int? = null,
+    val salary: Long? = null,
     val isExistSalary: Boolean? = null
 )
